pyproject.toml
[project]
name = "sample-mifos-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.12",
    "mcp>=1.6.0",
    "nest-asyncio>=1.6.0",
    "openai-agents>=0.0.8",
    "python-dotenv>=1.1.0",
]
